Create separate sub-modules in `node_extensions` for each type
