+++ crates/solidity-v2/outputs/cargo/ast/src/abi/node_extensions/contract_definition.rs
