+++ crates/solidity-v2/outputs/cargo/ast/src/abi/node_extensions/event_definition.rs
