+++ crates/solidity-v2/outputs/cargo/ast/src/abi/node_extensions/state_variable_definition.rs
