+++ crates/solidity-v2/outputs/cargo/ast/src/ast/node_extensions/mod.rs
@@ -1,7 +1,12 @@
-mod contracts;
-pub use contracts::ContractBase;
+mod contract_base;
+mod contract_definition;
+mod contract_members;
+mod interface_definition;
+pub use contract_base::ContractBase;
 
-mod expressions;
-mod functions;
-mod identifiers;
+mod function_call_expression;
+mod function_definition;
+mod identifier;
+mod identifier_path;
 mod source_unit;
+mod string_expression;
